Modificación de la interfaz y creación de la pantalla en landscape
