+++ app/src/main/res/layout-land/activity_main.xml
@@ -5,7 +5,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="110dp"
     tools:context="com.feedback.football.main.MainActivity">
 
     <LinearLayout
@@ -20,61 +19,91 @@
         android:paddingTop="@dimen/activity_vertical_margin">
 
         <!-- PARTIDO -->
-        <TextView
-            android:id="@+id/form_lbl_partidos"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            android:text="@string/lbl_partidos"
-            android:textAppearance="?android:attr/textAppearanceLarge"/>
-
-        <android.support.v7.widget.AppCompatSpinner
-            android:id="@+id/form_partidos_jornada"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:hint="@string/lbl_partidos"
-            android:inputType="text"
-            android:spinnerMode="dropdown"/>
+            android:orientation="horizontal"
+            android:weightSum="1">
+
+            <TextView
+                android:id="@+id/form_lbl_partidos"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.4"
+                android:text="@string/lbl_partidos"
+                android:textAlignment="center"
+                android:textAppearance="?android:attr/textAppearanceLarge"/>
+
+            <android.support.v7.widget.AppCompatSpinner
+                android:id="@+id/form_partidos_jornada"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.6"
+                android:hint="@string/lbl_partidos"
+                android:inputType="text"
+                android:spinnerMode="dropdown"
+                android:textAlignment="center"/>
+        </LinearLayout>
 
         <!-- MINUTO DE LA JUGADA-->
-        <TextView
-            android:id="@+id/form_minuto"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            android:text="@string/lbl_minuto"
-            android:textAppearance="?android:attr/textAppearanceLarge"/>
-
-        <android.support.v7.widget.AppCompatSpinner
-            android:id="@+id/form_jugada"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:entries="@array/minutos"
-            android:hint="@string/lbl_minuto"
-            android:inputType="text"
-            android:spinnerMode="dropdown"/>
+            android:orientation="horizontal"
+            android:weightSum="1">
+
+            <TextView
+                android:id="@+id/form_minuto"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.4"
+                android:text="@string/lbl_minuto"
+                android:textAlignment="center"
+                android:textAppearance="?android:attr/textAppearanceLarge"/>
+
+            <android.support.v7.widget.AppCompatSpinner
+                android:id="@+id/form_jugada"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.6"
+                android:entries="@array/minutos"
+                android:hint="@string/lbl_minuto"
+                android:inputType="text"
+                android:spinnerMode="dropdown"
+                android:textAlignment="center"/>
+        </LinearLayout>
+
 
         <!-- JUGADA -->
-        <TextView
-            android:id="@+id/form_lbl_jugada"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="10dp"
-            android:text="@string/lbl_jugada"
-            android:textAppearance="?android:attr/textAppearanceLarge"/>
-
-        <android.support.v7.widget.AppCompatSpinner
-            android:id="@+id/form_jugada"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginBottom="20dp"
-            android:entries="@array/jugadas"
-            android:hint="@string/lbl_jugada"
-            android:inputType="text"
-            android:spinnerMode="dropdown"/>
+            android:orientation="horizontal"
+            android:weightSum="1">
 
+            <TextView
+                android:id="@+id/form_lbl_jugada"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.4"
+                android:text="@string/lbl_jugada"
+                android:textAlignment="center"
+                android:textAppearance="?android:attr/textAppearanceLarge"/>
+
+            <android.support.v7.widget.AppCompatSpinner
+                android:id="@+id/form_jugada"
+                android:layout_width="0dp"
+                android:layout_height="wrap_content"
+                android:layout_weight="0.6"
+                android:entries="@array/jugadas"
+                android:hint="@string/lbl_jugada"
+                android:inputType="text"
+                android:spinnerMode="dropdown"
+                android:textAlignment="center"/>
+        </LinearLayout>
 
     </LinearLayout>
 
@@ -121,7 +150,7 @@
         android:id="@+id/editText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="90dp"
+        android:layout_marginBottom="50dp"
         android:inputType="text|textCapSentences"
         android:text="@string/lbl_arbitro"
         android:textAppearance="?android:attr/textAppearanceLarge"
